+++ src/models/services/tracks/ucscapi.rs
@@ -1,10 +1,12 @@
-use crate::models::services::tracks::helpers::get_prefered_track;
 use crate::{
     error::TGVError,
     models::{
         contig::Contig,
         reference::Reference,
         track::{Feature, Gene, Track},
+        services::{
+            tracks::TrackService
+        }
     },
 };
 use std::collections::HashMap;
@@ -15,6 +17,8 @@ use serde_json;
 pub struct UcscApiTrackService {
     reference: Reference,
     cached_tracks: HashMap<String, Track>,
+    
+    gene_name_lookup: HashMap<String, (String, usize)>, // gene_name -> (contig.full_name(), start)
 }
 
 impl UcscApiTrackService {
@@ -22,18 +26,29 @@ impl UcscApiTrackService {
         Self {
             reference,
             cached_tracks: HashMap::new(),
+            gene_name_lookup: HashMap::new(),
         }
     }
 
-    pub async fn load_contig(&mut self, contig: &Contig) -> Result<(), TGVError> {
+
+    /// Check if the contig's track is already cached. If not, load it.
+    /// Return true if loading is performed.
+    pub async fn check_or_load_contig(&mut self, contig: &Contig) -> Result<bool, TGVError> {
         if self.cached_tracks.contains_key(&contig.full_name()) {
-            return Ok(());
+            return Ok(false);
         }
 
         let track = self.query_feature_track(contig).await?;
         self.cached_tracks
             .insert(contig.full_name().to_string(), track);
-        Ok(())
+
+        // Populate gene_name_lookup
+        for gene in track.genes {
+            self.gene_name_lookup
+                .insert(gene.name.clone(), (contig.full_name().to_string(), gene.transcription_start));
+        }
+    
+        Ok(true)
     }
 
     async fn get_prefered_track(&self) -> Result<String, TGVError> {
@@ -63,7 +78,7 @@ impl UcscApiTrackService {
                     if track == "ncbiRefSeqSelect" {
                         prefered_track = Some(track.clone());
                         break;
-                    } else if (track == "ncbiRefSeqCurated") {
+                    } else if track == "ncbiRefSeqCurated" {
                         prefered_track = Some(track.clone());
                     } else if track == "ncbiRefSeq"
                         && prefered_track != Some("ncbiRefSeqCurated".to_string())
@@ -87,4 +102,13 @@ impl UcscApiTrackService {
             )),
         }
     }
+}
+
+impl TrackService for UcscApiTrackService {
+    async fn query_feature_track(&self, contig: &Contig) -> Result<Track, TGVError> {
+        let track = self.query_feature_track(contig).await?;
+        Ok(track)
+    }
+
+    async fn 
 }
